Tell git to ignore these files
The `.gitignore` file is a list of file patterns
that Git will look at: Git will ignore (not track)
any file(s) matching these patterns.

We don't want git to track (version) the `node_modules/`
directory because it contains lots of files that we didn't
write and which we can easily regenerate by running `npm i`.